Rename test files to get around Windows limitation
Having a file called "a1" and a folder called "A1" screws with
filesystems where names are case insensitive. In general, this
could actually be a problem if there's a file named the same
as one of the target folders. It's just a bit more obvious on
Windows when the casing isn't taken into account.

Opened #4